fastapi-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pydantic
rapidfuzz
sentence-transformers
faiss-cpu
numpy
